git-svn-id: http://svn.python.org/projects/python/branches/ssize_t@42380 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Objects/unicodeobject.c
@@ -6535,7 +6535,7 @@ unicode_buffer_getwritebuf(PyUnicodeObject *self, Py_ssize_t index,
     return -1;
 }
 
-static Py_ssize_t
+static int
 unicode_buffer_getsegcount(PyUnicodeObject *self,
 			   Py_ssize_t *lenp)
 {
@@ -6544,7 +6544,7 @@ unicode_buffer_getsegcount(PyUnicodeObject *self,
     return 1;
 }
 
-static Py_ssize_t
+static int
 unicode_buffer_getcharbuf(PyUnicodeObject *self,
 			  Py_ssize_t index,
 			  const void **ptr)
